refactor: room 패키지 수정
