+++ src/main/java/com/travel/role/domain/room/dto/response/InviteResponseDTO.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.dto;
+package com.travel.role.domain.room.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class MemberDTO {
-	private String name;
-	private String profile;
+public class InviteResponseDTO {
+	private Long roomId;
 }
