+++ src/main/java/com/travel/role/domain/room/dto/response/MemberDTO.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.dto;
+package com.travel.role.domain.room.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class InviteResponseDTO {
-	private Long roomId;
+public class MemberDTO {
+	private String name;
+	private String profile;
 }
