+++ src/main/java/com/travel/role/domain/room/repository/RoomRepository.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.travel.role.domain.room.entity.Room;
+import com.travel.role.domain.room.repository.querydsl.RoomQuerydsl;
 
 public interface RoomRepository extends JpaRepository<Room, Long>, RoomQuerydsl {
     boolean existsByRoomInviteCode(String inviteCode);
