+++ src/main/java/com/travel/role/domain/room/service/RoomService.java
@@ -22,10 +22,10 @@ import com.travel.role.domain.room.entity.ParticipantRole;
 import com.travel.role.domain.room.entity.Room;
 import com.travel.role.domain.room.entity.RoomParticipant;
 import com.travel.role.domain.room.entity.RoomRole;
-import com.travel.role.domain.room.dto.InviteResponseDTO;
-import com.travel.role.domain.room.dto.MakeRoomRequestDTO;
-import com.travel.role.domain.room.dto.MemberDTO;
-import com.travel.role.domain.room.dto.RoomResponseDTO;
+import com.travel.role.domain.room.dto.response.InviteResponseDTO;
+import com.travel.role.domain.room.dto.request.MakeRoomRequestDTO;
+import com.travel.role.domain.room.dto.response.MemberDTO;
+import com.travel.role.domain.room.dto.response.RoomResponseDTO;
 import com.travel.role.global.exception.room.AlreadyExistInRoomException;
 import com.travel.role.global.exception.room.InvalidInviteCode;
 import com.travel.role.global.exception.room.InvalidLocalDateException;
