+++ src/test/java/com/travel/role/domain/room/RoomTest.java
@@ -16,9 +16,9 @@ import com.travel.role.domain.room.repository.RoomParticipantRepository;
 import com.travel.role.domain.room.repository.RoomRepository;
 import com.travel.role.domain.room.entity.Room;
 import com.travel.role.domain.room.entity.RoomParticipant;
-import com.travel.role.domain.room.dto.MakeRoomRequestDTO;
+import com.travel.role.domain.room.dto.request.MakeRoomRequestDTO;
 import com.travel.role.domain.room.repository.ParticipantRoleRepository;
-import com.travel.role.domain.room.dto.RoomResponseDTO;
+import com.travel.role.domain.room.dto.response.RoomResponseDTO;
 import com.travel.role.domain.room.service.RoomService;
 import com.travel.role.domain.user.repository.UserRepository;
 import com.travel.role.domain.user.entity.User;
